more code review comments
