+++ api/fees/types.go
@@ -6,29 +6,9 @@
 package fees
 
 import (
-	"sync"
-
 	"github.com/ethereum/go-ethereum/common/hexutil"
-	"github.com/vechain/thor/v2/bft"
-	"github.com/vechain/thor/v2/cache"
-	"github.com/vechain/thor/v2/chain"
 )
 
-type Fees struct {
-	data *FeesData
-	done chan struct{}
-	wg   sync.WaitGroup
-}
-type FeeCacheEntry struct {
-	baseFee      *hexutil.Big
-	gasUsedRatio float64
-}
-type FeesData struct {
-	repo           *chain.Repository
-	cache          *cache.PrioCache
-	bft            bft.Committer
-	backtraceLimit uint32 // The max number of blocks to backtrace.
-}
 type FeesHistory struct {
 	OldestBlock   *uint32        `json:"oldestBlock"`
 	BaseFees      []*hexutil.Big `json:"baseFees"`
